requirements.txt
simplejson
pyaml
testtools
nose
nose_xunitmp
requests
selenium
splinter
qubell-api-python-client